// client/src/app/models/capacitation.ts
export interface CapacitationInterface{
    id_capacitacion?: number;
    id_administrador?: number;
    nombre?: string;
    descripcion?: string;
    costo?: number;
    duracion?: string;
    expositor?: string;
    fecha_inicio?: Date;
    fecha_fin?: Date;
    fecha_inicio_insc?: Date;
    fecha_fin_insc?: Date;
    imagen?: string;
    lugar?: string;
    hora?: string;
    pathImagen?: string;
} 